Implemented threadsafe random, checksum
I changed the random number generation to be threadsafe in that it won’t be the same number over and over and over.

Furthermore, I implemented a Fletcher 64 checksuming algorithm to generate a checksum of the data that’s in a record.

Next step: Write the record to the file.