Cleanup and implement multi-threaded writes
First, I moved all the globals, includes, definitions, etc., to a main.h.
Commented out unneeded screen output for now.
Added timestamp (using nanoseconds since epoch) to the records which is made at the beginning of a record being created.
Switched to defined record indexes.
This should be all the required parts of part 2, with some extras. A nice checksum, a timestamp, etc.
Time to move on to part 4.