Updating record to have a head & tail checksum
Because systems could write the last byte first or the first byte first, I put a checksum at either side to be able to redundantly check for data.

While doing this, I updated the Fletcher64 to be simpler to call. Now you only need to provide the record and it correctly checks record_indexes.h so that it covers the correct ranges automatically. This should be the last change to Fletcher64.

Finally, I added a debug flag to the checker program so that it can output information about bad records.