Change pwrite to allow for partial writes
According to the POSIX specifications, pwrite is “automic” in that a kill signal will not interrupt a program until a pwrite is completed. As such, short of managing an absolute power kill (which is outside the scope of this assignment), I need to differientiate the way I handle the pwrites. I chose to write each long individually, but this may not be ideal as the number of writes is 1024 per record. The only saving grace is that the file is only opened once and referenced from then.

Still, this does give me partial writes… I assume this will work for this assignment.