Extending the checker for better statistics
I am now comparing timestamps to the actual writes. Because of random write location, a thread could write to a later section in the file earlier and an earlier section in the file later and thus the status of the last write could be unsuccessful.

I am tracking the hard numbers for good and bad writes. From here, I will extend the program to show percentages of successful writes.