Removing mutexes
Given the style of what I’m doing…I don’t need a mutex in this case. The only way it makes sense is if I had a mutex for each possible record, but the overhead of such an operation is extreme and not worth it. For the record I have a maximum of 100 possible threads and 16,384 possible records that could be randomly selected over a uniform distribution. I’ll play these odds.