Incorporating output into run_experiments.sh
Since file output was added to the checker, I added it to the runner as well.

Also saw a bug that shouldn’t be a problem but I fixed anyway, that of asking for an argument with the help flag (-?). I wasn’t parsing for one, so it shouldn’t have been an issue but just in case…